Rename component for street modes
