+++ ui/src/lib/components/ui/StreetModes.svelte
